Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
openai==0.28.1
httpx==0.24.1
PyGithub==1.59.1
boto3==1.34.0
botocore==1.34.0
azure-mgmt-resource==23.0.1
azure-identity==1.15.0
google-cloud-resource-manager==1.10.3
google-cloud-storage==2.10.0
pydantic==2.5.0 